Move plookup tables into proof_system
