+++ cpp/src/barretenberg/proof_system/CMakeLists.txt
@@ -1,1 +1,1 @@
-barretenberg_module(proof_system polynomials crypto_generators plonk)
+barretenberg_module(proof_system polynomials crypto_generators crypto_pedersen_hash)
