+++ cpp/src/barretenberg/proof_system/plookup_tables/keccak/keccak_output.hpp
