+++ cpp/src/barretenberg/stdlib/commitment/pedersen/pedersen_plookup.cpp
@@ -3,7 +3,7 @@
 #include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
 #include "../../hash/pedersen/pedersen_plookup.hpp"
 
-#include "barretenberg/plonk/composer/plookup_tables/types.hpp"
+#include "barretenberg/proof_system/plookup_tables/types.hpp"
 #include "../../primitives/composers/composers.hpp"
 #include "../../primitives/plookup/plookup.hpp"
