+++ cpp/src/barretenberg/stdlib/hash/blake2s/blake2s_plookup.cpp
@@ -1,8 +1,8 @@
 #include "blake2s_plookup.hpp"
 #include "blake_util.hpp"
 
-#include "barretenberg/plonk/composer/plookup_tables/plookup_tables.hpp"
-#include "barretenberg/plonk/composer/plookup_tables/sha256.hpp"
+#include "barretenberg/proof_system/plookup_tables/plookup_tables.hpp"
+#include "barretenberg/proof_system/plookup_tables/sha256.hpp"
 #include "barretenberg/plonk/composer/ultra_composer.hpp"
 #include "barretenberg/stdlib/primitives/bit_array/bit_array.hpp"
 #include "barretenberg/stdlib/primitives/field/field.hpp"
