+++ cpp/src/barretenberg/stdlib/hash/blake3s/blake3s_plookup.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <array>
-#include "barretenberg/plonk/composer/plookup_tables/plookup_tables.hpp"
+#include "barretenberg/proof_system/plookup_tables/plookup_tables.hpp"
 #include "barretenberg/stdlib/primitives/uint/uint.hpp"
 
 #include "barretenberg/numeric/bitop/sparse_form.hpp"
