+++ cpp/src/barretenberg/stdlib/hash/sha256/sha256_plookup.cpp
@@ -1,7 +1,7 @@
 #include "sha256_plookup.hpp"
 
-#include "barretenberg/plonk/composer/plookup_tables/plookup_tables.hpp"
-#include "barretenberg/plonk/composer/plookup_tables/sha256.hpp"
+#include "barretenberg/proof_system/plookup_tables/plookup_tables.hpp"
+#include "barretenberg/proof_system/plookup_tables/sha256.hpp"
 #include "barretenberg/plonk/composer/ultra_composer.hpp"
 #include "barretenberg/stdlib/primitives/bit_array/bit_array.hpp"
 #include "barretenberg/stdlib/primitives/field/field.hpp"
