+++ cpp/src/barretenberg/stdlib/primitives/plookup/plookup.hpp
@@ -1,9 +1,9 @@
 #pragma once
 #include <array>
 #include <vector>
-#include "barretenberg/plonk/composer/plookup_tables/plookup_tables.hpp"
+#include "barretenberg/proof_system/plookup_tables/plookup_tables.hpp"
 #include "barretenberg/plonk/composer/ultra_composer.hpp"
-#include "barretenberg/plonk/composer/plookup_tables/types.hpp"
+#include "barretenberg/proof_system/plookup_tables/types.hpp"
 #include "barretenberg/stdlib/primitives/field/field.hpp"
 
 namespace proof_system::plonk {
